Refactor API modules and add basic tests
