+++ AGENTS.md
@@ -17,6 +17,9 @@
 | modules/auth.ts         | Функции аутентификации пользователей |
 | modules/telegram.ts     | Интеграция Telegram (токены, поиск) |
 | modules/subscription.ts | Запросы и статистика подписок |
+| modules/admin.ts        | Проверки прав админа и сообщения |
+| modules/payments.ts     | Оплата подписок и транзакции |
+| modules/offers.ts       | Управление рекламными предложениями |
 | schema.prisma         | Схема базы данных Prisma                                            |
 | theme.css             | Основные стили интерфейса                                           |
 | PRD.md                | Продуктовые требования, описание сценариев использования            |
@@ -50,6 +53,13 @@
 ### [2025-06-06]
 - [СДЕЛАНО] Вынесены функции из api.ts в модули auth.ts, telegram.ts и subscription.ts для лучшей модульности
 - [НОВАЯ ЗАДАЧА] Перенести остальные API-функции в отдельные файлы
+
+### [2025-06-06]
+- [СДЕЛАНО] Удалены неиспользуемые импорты и TOKEN_EXPIRY_HOURS в api.ts
+- [СДЕЛАНО] Созданы модули admin.ts, payments.ts и offers.ts
+- [СДЕЛАНО] Проверки администратора перенесены в requireAdmin
+- [СДЕЛАНО] В ключевых платежных функциях добавлены Prisma-транзакции
+- [СДЕЛАНО] Добавлены простые юнит‑тесты для auth, subscription и telegram
 
 ---
