+++ api.ts
@@ -1,15 +1,11 @@
 import { db } from "~/server/db";
 import { getAuth, sendEmail } from "~/server/actions";
-import { z } from "zod";
-import { nanoid } from "nanoid";
 import crypto from "crypto";
 import axios from "axios";
-import fs from "fs";
 import util from "util";
 import { exec as execCallback } from "child_process";
 
 // Constants for authentication
-const TOKEN_EXPIRY_HOURS = 24;
 const EMAIL_VERIFICATION_EXPIRY_HOURS = 48;
 const PASSWORD_RESET_EXPIRY_HOURS = 24;
 
@@ -19,6 +15,16 @@ const exec = util.promisify(execCallback);
 export { getCurrentUser } from "./modules/auth";
 export { getUserSubscriptions, getSubscriptionStats } from "./modules/subscription";
 export { generateTelegramToken, verifyTelegramToken, getUserByTelegramId } from "./modules/telegram";
+export { requireAdmin, getUsers, sendAdminMessage } from "./modules/admin";
+export {
+  initiatePayment,
+  approvePayment,
+  rejectPayment,
+  getAllPayments,
+  extendSubscription,
+  getUserPayments,
+} from "./modules/payments";
+export { createOffer, updateOffer, deleteOffer, getAllOffers } from "./modules/offers";
 
 
 // Subscription Plans
@@ -62,330 +68,8 @@ export async function createSubscriptionPlan(input: {
 }
 
 // Payment System
-export async function initiatePayment(input: {
-  planId: string;
-  paymentDetails?: string;
-}) {
-  try {
-    const auth = await getAuth({ required: true });
-
-    // Get the subscription plan
-    const plan = await db.subscriptionPlan.findUnique({
-      where: { id: input.planId },
-    });
-
-    if (!plan) {
-      throw new Error("Subscription plan not found");
-    }
-
-    // Calculate end date based on plan duration
-    const startDate = new Date();
-    const endDate = new Date(startDate);
-
-    switch (plan.duration) {
-      case "7days":
-        endDate.setDate(endDate.getDate() + 7);
-        break;
-      case "1month":
-        endDate.setMonth(endDate.getMonth() + 1);
-        break;
-      case "3months":
-        endDate.setMonth(endDate.getMonth() + 3);
-        break;
-      case "6months":
-        endDate.setMonth(endDate.getMonth() + 6);
-        break;
-      case "12months":
-        endDate.setMonth(endDate.getMonth() + 12);
-        break;
-      default:
-        endDate.setMonth(endDate.getMonth() + 1); // Default to 1 month
-    }
-
-    // Create payment record
-    const payment = await db.payment.create({
-      data: {
-        amount: plan.price,
-        userId: auth.userId,
-        paymentDetails: input.paymentDetails,
-      },
-    });
-
-    // Create pending subscription
-    const subscription = await db.subscription.create({
-      data: {
-        planName: plan.name,
-        planDuration: plan.duration,
-        price: plan.price,
-        startDate,
-        endDate,
-        userId: auth.userId,
-        paymentId: payment.id,
-      },
-    });
-
-    // Notify admins about new payment
-    const admins = await db.user.findMany({ where: { isAdmin: true } });
-    for (const admin of admins) {
-      if (admin.telegramId) {
-        // In a real implementation, you would send Telegram notification here
-        console.log(
-          `Notify admin ${admin.id} via Telegram about payment ${payment.id}`,
-        );
-      }
-
-      // Send email notification if we have an admin
-      if (admin.id) {
-        try {
-          await sendEmail({
-            toUserId: admin.id,
-            subject: "New VPN Subscription Payment",
-            markdown: `
-### New Payment Requires Approval
-
-**User ID:** ${auth.userId}
-**Plan:** ${plan.name}
-**Amount:** ${plan.price}
-**Duration:** ${plan.duration}
-
-Please review and approve this payment in the admin dashboard.
-            `,
-          });
-        } catch (error) {
-          console.error("Failed to send admin notification email:", error);
-        }
-      }
-    }
-
-    return { paymentId: payment.id, subscriptionId: subscription.id };
-  } catch (error) {
-    console.error("Error in initiatePayment:", error);
-    throw new Error("Failed to initiate payment");
-  }
-}
-
-export async function approvePayment(input: { paymentId: string }) {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    // Update payment status
-    const payment = await db.payment.update({
-      where: { id: input.paymentId },
-      data: {
-        status: "completed",
-        adminApproved: true,
-        adminId: auth.userId,
-      },
-      include: { subscription: true },
-    });
-
-    if (!payment.subscription) {
-      throw new Error("No subscription found for this payment");
-    }
-
-    // Activate the subscription
-    const subscription = await db.subscription.update({
-      where: { id: payment.subscription.id },
-      data: { status: "active" },
-      include: { user: true },
-    });
-
-    // Notify the user that their subscription is now active
-    if (subscription.user.id) {
-      try {
-        await sendEmail({
-          toUserId: subscription.user.id,
-          subject: "Your VPN Subscription is Active",
-          markdown: `
-### Your VPN Subscription is Now Active!
-
-**Plan:** ${subscription.planName}
-**Duration:** ${subscription.planDuration}
-**Valid Until:** ${new Date(subscription.endDate).toLocaleDateString()}
-
-Thank you for your purchase. You can now access all VPN features.
-          `,
-        });
-      } catch (error) {
-        console.error("Failed to send subscription activation email:", error);
-      }
-    }
-
-    return { success: true, subscription };
-  } catch (error) {
-    console.error("Error in approvePayment:", error);
-    throw new Error("Failed to approve payment");
-  }
-}
-
-export async function rejectPayment(input: {
-  paymentId: string;
-  reason?: string;
-}) {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    // Update payment status
-    const payment = await db.payment.update({
-      where: { id: input.paymentId },
-      data: {
-        status: "cancelled",
-        adminApproved: false,
-        adminId: auth.userId,
-      },
-      include: { subscription: true, user: true },
-    });
-
-    if (payment.subscription) {
-      // Cancel the subscription
-      await db.subscription.update({
-        where: { id: payment.subscription.id },
-        data: { status: "suspended" },
-      });
-    }
-
-    // Notify the user that their payment was rejected
-    if (payment.user.id) {
-      try {
-        await sendEmail({
-          toUserId: payment.user.id,
-          subject: "Your VPN Payment Was Rejected",
-          markdown: `
-### Your VPN Payment Was Rejected
-
-**Amount:** ${payment.amount}
-**Reason:** ${input.reason || "No reason provided"}
-
-Please contact support for more information or try again with a different payment method.
-          `,
-        });
-      } catch (error) {
-        console.error("Failed to send payment rejection email:", error);
-      }
-    }
-
-    return { success: true };
-  } catch (error) {
-    console.error("Error in rejectPayment:", error);
-    throw new Error("Failed to reject payment");
-  }
-}
 
 // User Management API
-export async function getUsers(input?: { role?: string }) {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    // Build the where clause based on the role filter
-    const where = input?.role ? { role: input.role } : {};
-
-    const users = await db.user.findMany({
-      where,
-      select: {
-        id: true,
-        name: true,
-        role: true,
-        telegramUsername: true,
-        createdAt: true,
-        isAdmin: true,
-        _count: {
-          select: {
-            subscriptions: true,
-            payments: true,
-          },
-        },
-      },
-      orderBy: { createdAt: "desc" },
-    });
-
-    return users;
-  } catch (error) {
-    console.error("Error in getUsers:", error);
-    throw new Error("Failed to get users");
-  }
-}
-
-export async function sendAdminMessage(input: {
-  content: string;
-  targetRole?: string;
-  recipientIds?: string[];
-}) {
-  try {
-    const auth = await getAuth({ required: true });
-    const sender = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!sender?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    // If targetRole is provided, send to all users with that role
-    if (input.targetRole) {
-      // Only allow sending to VIP and tester roles as specified in requirements
-      if (input.targetRole !== "VIP" && input.targetRole !== "tester") {
-        throw new Error(
-          "Can only send role-based messages to VIP and tester users",
-        );
-      }
-
-      const targetUsers = await db.user.findMany({
-        where: { role: input.targetRole },
-        select: { id: true },
-      });
-
-      const messages = await Promise.all(
-        targetUsers.map((user) =>
-          db.message.create({
-            data: {
-              content: input.content,
-              targetRole: input.targetRole,
-              senderId: auth.userId,
-              recipientId: user.id,
-            },
-          }),
-        ),
-      );
-
-      return { success: true, messageCount: messages.length };
-    }
-
-    // If recipientIds is provided, send to those specific users
-    else if (input.recipientIds && input.recipientIds.length > 0) {
-      const messages = await Promise.all(
-        input.recipientIds.map((recipientId) =>
-          db.message.create({
-            data: {
-              content: input.content,
-              senderId: auth.userId,
-              recipientId,
-            },
-          }),
-        ),
-      );
-
-      return { success: true, messageCount: messages.length };
-    }
-
-    throw new Error("Either targetRole or recipientIds must be provided");
-  } catch (error) {
-    console.error("Error in sendAdminMessage:", error);
-    throw new Error("Failed to send admin message");
-  }
-}
 
 export async function setDiscount(input: {
   code: string;
@@ -790,94 +474,6 @@ export async function getDiscountStats() {
   }
 }
 
-export async function getAllPayments() {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    return await db.payment.findMany({
-      include: { user: true, subscription: true },
-      orderBy: { createdAt: "desc" },
-    });
-  } catch (error) {
-    console.error("Error in getAllPayments:", error);
-    return [];
-  }
-}
-
-export async function extendSubscription(input: {
-  subscriptionId: string;
-  months: number;
-}) {
-  try {
-    const auth = await getAuth({ required: true });
-
-    // Check if admin or subscription owner
-    const subscription = await db.subscription.findUnique({
-      where: { id: input.subscriptionId },
-      include: { user: true },
-    });
-
-    if (!subscription) {
-      throw new Error("Subscription not found");
-    }
-
-    const isAdmin = await db.user.findUnique({
-      where: { id: auth.userId },
-      select: { isAdmin: true },
-    });
-
-    // Only allow admin or subscription owner to extend
-    if (subscription.userId !== auth.userId && !isAdmin?.isAdmin) {
-      throw new Error("Unauthorized: You cannot extend this subscription");
-    }
-
-    // If not admin, require payment process
-    if (!isAdmin?.isAdmin) {
-      throw new Error(
-        "Please use the payment system to extend your subscription",
-      );
-    }
-
-    // Calculate new end date
-    const newEndDate = new Date(subscription.endDate);
-    newEndDate.setMonth(newEndDate.getMonth() + input.months);
-
-    // Update subscription
-    const updatedSubscription = await db.subscription.update({
-      where: { id: input.subscriptionId },
-      data: {
-        endDate: newEndDate,
-        status: "active", // Ensure it's active after extension
-      },
-    });
-
-    return updatedSubscription;
-  } catch (error) {
-    console.error("Error in extendSubscription:", error);
-    throw new Error("Failed to extend subscription");
-  }
-}
-
-// User Subscription Management
-export async function getUserPayments() {
-  try {
-    const auth = await getAuth({ required: true });
-
-    return await db.payment.findMany({
-      where: { userId: auth.userId },
-      include: { subscription: true },
-      orderBy: { createdAt: "desc" },
-    });
-  } catch (error) {
-    console.error("Error in getUserPayments:", error);
-    return [];
-  }
-}
 
 // Xray API Management
 
@@ -1473,143 +1069,6 @@ export async function _seedNewsAndVpnOptions() {
   }
 }
 
-// Offer Management
-
-export async function createOffer(input: {
-  title: string;
-  description: string;
-  imageUrl?: string;
-  price: number;
-  isVisible?: boolean;
-  priority?: number;
-  startDate?: string;
-  endDate?: string;
-}) {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    return await db.offer.create({
-      data: {
-        title: input.title,
-        description: input.description,
-        imageUrl: input.imageUrl,
-        price: input.price,
-        isVisible: input.isVisible !== undefined ? input.isVisible : true,
-        priority: input.priority || 0,
-        startDate: input.startDate ? new Date(input.startDate) : null,
-        endDate: input.endDate ? new Date(input.endDate) : null,
-      },
-    });
-  } catch (error) {
-    console.error("Error in createOffer:", error);
-    throw new Error("Failed to create offer");
-  }
-}
-
-export async function updateOffer(input: {
-  id: string;
-  title?: string;
-  description?: string;
-  imageUrl?: string;
-  price?: number;
-  isVisible?: boolean;
-  priority?: number;
-  startDate?: string | null;
-  endDate?: string | null;
-}) {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    // Prepare update data
-    const updateData: any = {};
-    if (input.title !== undefined) updateData.title = input.title;
-    if (input.description !== undefined) updateData.description = input.description;
-    if (input.imageUrl !== undefined) updateData.imageUrl = input.imageUrl;
-    if (input.price !== undefined) updateData.price = input.price;
-    if (input.isVisible !== undefined) updateData.isVisible = input.isVisible;
-    if (input.priority !== undefined) updateData.priority = input.priority;
-    if (input.startDate !== undefined) updateData.startDate = input.startDate ? new Date(input.startDate) : null;
-    if (input.endDate !== undefined) updateData.endDate = input.endDate ? new Date(input.endDate) : null;
-
-    return await db.offer.update({
-      where: { id: input.id },
-      data: updateData,
-    });
-  } catch (error) {
-    console.error("Error in updateOffer:", error);
-    throw new Error("Failed to update offer");
-  }
-}
-
-export async function deleteOffer(input: { id: string }) {
-  try {
-    const auth = await getAuth({ required: true });
-    const user = await db.user.findUnique({ where: { id: auth.userId } });
-
-    if (!user?.isAdmin) {
-      throw new Error("Unauthorized: Admin access required");
-    }
-
-    await db.offer.delete({
-      where: { id: input.id },
-    });
-
-    return { success: true };
-  } catch (error) {
-    console.error("Error in deleteOffer:", error);
-    throw new Error("Failed to delete offer");
-  }
-}
-
-export async function getAllOffers(input?: { includeHidden?: boolean }) {
-  try {
-    const auth = await getAuth();
-    const isAdmin = auth.status === "authenticated" ? 
-      await db.user.findUnique({ where: { id: auth.userId }, select: { isAdmin: true } }) : 
-      null;
-
-    // Non-admins can only see visible offers
-    let where: any = !isAdmin?.isAdmin && !input?.includeHidden ? 
-      { isVisible: true } : 
-      {};
-
-    // Add date filtering for non-admins
-    if (!isAdmin?.isAdmin) {
-      const now = new Date();
-      where.AND = [
-        { OR: [
-          { startDate: null },
-          { startDate: { lte: now } }
-        ]},
-        { OR: [
-          { endDate: null },
-          { endDate: { gte: now } }
-        ]}
-      ];
-    }
-
-    return await db.offer.findMany({
-      where,
-      orderBy: [
-        { priority: "desc" },
-        { createdAt: "desc" }
-      ],
-    });
-  } catch (error) {
-    console.error("Error in getAllOffers:", error);
-    return [];
-  }
-}
 
 // Email Authentication Preparation
 // These functions will need to be implemented with your email service
